Make new file _list_pool.py
Use that for pool listing functions. _pool.py was getting too big.
